liualgotrader/requirements/dev.txt
iso8601==1.0.2
requests==2.27.1
deprecated==1.2.13
urllib3==1.26.8
numpy==1.22.3
mypy==0.940
black==22.1.0
pylint==2.12.2
flake8==4.0.1
Cython==0.29.28
alpaca-trade-api==1.5.1
google-cloud-logging==3.0.0
therapist==2.2.0
jupyter==1.0.0
asyncpg==0.25.0
pygit2==1.9.0
pytest==7.1.0
pytest-asyncio==0.18.2
pytest-cov==3.0.0
nb-clean==2.0.2
matplotlib==3.5.1
tabulate==0.8.9
sphinx==4.4.0
sphinx-autodoc-typehints==1.17.0
setuptools==60.9.3
wheel==0.37.1
twine==3.8.0
streamlit==1.7.0
stockstats==0.4.1
nest_asyncio==1.5.4
nbqa==1.3.1
pre-commit==2.17.0
empyrical==0.5.5
sklearn==0.0
plotly==5.6.0
hypothesis==6.39.3
torch==1.11.0
parsedatetime==2.6
scipy==1.8.0
polygon-api-client==0.1.11
bandit==1.7.4
finnhub-python==2.4.10
pandas==1.4.1
quantstats==0.0.50
tqdm==4.63.0
mnqueues==0.0.37
toml==0.10.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-serializinghtml==1.1.5
fire==0.4.0
opentelemetry-exporter-gcp-trace==1.0.0
websocket-client==1.3.1
types-requests==2.27.12
types-pytz==2021.3.5
pandas_market_calendars==3.4
psutil==5.9.0
types-python-dateutil==2.8.9
autoflake
